refactor: encapsulate login logic and reuse it in `LoginModal` and `Login` page
