+++ projects/app/src/pages/login/index.tsx
@@ -1,67 +1,19 @@
-import React, { useState, useCallback, useEffect, useMemo } from 'react';
-import {
-  Box,
-  Button,
-  Center,
-  Drawer,
-  DrawerCloseButton,
-  DrawerContent,
-  DrawerOverlay,
-  Flex,
-  useDisclosure
-} from '@chakra-ui/react';
-import { LoginPageTypeEnum } from '@/web/support/user/login/constants';
-import { useSystemStore } from '@/web/common/system/useSystemStore';
+import React, { useCallback } from 'react';
+import { Flex } from '@chakra-ui/react';
 import type { ResLogin } from '@/global/support/api/userRes.d';
 import { useRouter } from 'next/router';
-import { useUserStore } from '@/web/support/user/useUserStore';
-import { useChatStore } from '@/web/core/chat/context/useChatStore';
-import dynamic from 'next/dynamic';
 import { serviceSideProps } from '@/web/common/i18n/utils';
 import { clearToken } from '@/web/support/user/auth';
-import Script from 'next/script';
-import Loading from '@fastgpt/web/components/common/MyLoading';
-import { useLocalStorageState, useMount } from 'ahooks';
-import { useTranslation } from 'next-i18next';
-import I18nLngSelector from '@/components/Select/I18nLngSelector';
-import { useSystem } from '@fastgpt/web/hooks/useSystem';
-import { GET } from '@/web/common/api/request';
-import { getDocPath } from '@/web/common/system/doc';
+import { useMount } from 'ahooks';
 import { getWebReqUrl } from '@fastgpt/web/common/system/utils';
-import LoginForm from '@/pageComponents/login/LoginForm/LoginForm';
-import { getBdVId } from '@/web/support/marketing/utils';
-
-const RegisterForm = dynamic(() => import('@/pageComponents/login/RegisterForm'));
-const ForgetPasswordForm = dynamic(() => import('@/pageComponents/login/ForgetPasswordForm'));
-const WechatForm = dynamic(() => import('@/pageComponents/login/LoginForm/WechatForm'));
-const CommunityModal = dynamic(() => import('@/components/CommunityModal'));
-
-const ipDetectURL = 'https://qifu-api.baidubce.com/ip/local/geo/v1/district';
+import { LoginContainer } from '@/pageComponents/login';
 
 const Login = ({ ChineseRedirectUrl }: { ChineseRedirectUrl: string }) => {
   const router = useRouter();
-  const { t } = useTranslation();
   const { lastRoute = '' } = router.query as { lastRoute: string };
-  const { feConfigs } = useSystemStore();
-  const [pageType, setPageType] = useState<`${LoginPageTypeEnum}`>(LoginPageTypeEnum.passwordLogin);
-  const { setUserInfo } = useUserStore();
-  const { setLastChatAppId } = useChatStore();
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const { isPc } = useSystem();
-
-  const {
-    isOpen: isOpenCookiesDrawer,
-    onOpen: onOpenCookiesDrawer,
-    onClose: onCloseCookiesDrawer
-  } = useDisclosure();
-  const cookieVersion = '1';
-  const [localCookieVersion, setLocalCookieVersion] =
-    useLocalStorageState<string>('localCookieVersion');
 
   const loginSuccess = useCallback(
     (res: ResLogin) => {
-      setUserInfo(res.user);
-
       const decodeLastRoute = decodeURIComponent(lastRoute);
 
       const navigateTo =
@@ -71,224 +23,45 @@ const Login = ({ ChineseRedirectUrl }: { ChineseRedirectUrl: string }) => {
 
       router.push(navigateTo);
     },
-    [setUserInfo, lastRoute, router]
+    [lastRoute, router]
   );
-
-  const DynamicComponent = useMemo(() => {
-    const TypeMap = {
-      [LoginPageTypeEnum.passwordLogin]: LoginForm,
-      [LoginPageTypeEnum.register]: RegisterForm,
-      [LoginPageTypeEnum.forgetPassword]: ForgetPasswordForm,
-      [LoginPageTypeEnum.wechat]: WechatForm
-    };
-
-    // @ts-ignore
-    const Component = TypeMap[pageType];
-
-    return <Component setPageType={setPageType} loginSuccess={loginSuccess} />;
-  }, [pageType, loginSuccess]);
-
-  /* default login type */
-  useEffect(() => {
-    //delete bd_vid and show WeChat login
-    setPageType(
-      feConfigs?.oauth?.wechat ? LoginPageTypeEnum.wechat : LoginPageTypeEnum.passwordLogin
-    );
-    // init store
-    setLastChatAppId('');
-  }, [feConfigs?.oauth, setLastChatAppId]);
-
-  const {
-    isOpen: isOpenRedirect,
-    onOpen: onOpenRedirect,
-    onClose: onCloseRedirect
-  } = useDisclosure();
-  const [showRedirect, setShowRedirect] = useLocalStorageState<boolean>('showRedirect', {
-    defaultValue: true
-  });
-  const checkIpInChina = useCallback(async () => {
-    try {
-      const res = await GET<any>(ipDetectURL);
-      const country = res?.country;
-      if (
-        country &&
-        country === '中国' &&
-        res.prov !== '中国香港' &&
-        res.prov !== '中国澳门' &&
-        res.prov !== '中国台湾'
-      ) {
-        onOpenRedirect();
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  }, [onOpenRedirect]);
 
   useMount(() => {
     clearToken();
     router.prefetch('/dashboard/apps');
-
-    ChineseRedirectUrl && showRedirect && checkIpInChina();
-    localCookieVersion !== cookieVersion && onOpenCookiesDrawer();
   });
 
   return (
-    <>
-      {feConfigs.googleClientVerKey && (
-        <Script
-          src={`https://www.recaptcha.net/recaptcha/api.js?render=${feConfigs.googleClientVerKey}`}
-        ></Script>
-      )}
-
+    <Flex
+      alignItems={'center'}
+      justifyContent={'center'}
+      bg={`url(${getWebReqUrl('/icon/login-bg.svg')}) no-repeat`}
+      backgroundSize={'cover'}
+      userSelect={'none'}
+      h={'100%'}
+    >
       <Flex
-        alignItems={'center'}
-        justifyContent={'center'}
-        bg={`url(${getWebReqUrl('/icon/login-bg.svg')}) no-repeat`}
-        backgroundSize={'cover'}
-        userSelect={'none'}
-        h={'100%'}
+        flexDirection={'column'}
+        w={['100%', '556px']}
+        h={['100%', '677px']}
+        bg={'white'}
+        px={['5vw', '88px']}
+        py={['5vh', '64px']}
+        borderRadius={[0, '16px']}
+        boxShadow={[
+          '',
+          '0px 32px 64px -12px rgba(19, 51, 107, 0.20), 0px 0px 1px 0px rgba(19, 51, 107, 0.20)'
+        ]}
       >
-        {isPc && (
-          <Box position={'absolute'} top={'24px'} right={'50px'}>
-            <I18nLngSelector />
-          </Box>
-        )}
-        <Flex
-          flexDirection={'column'}
-          w={['100%', '556px']}
-          h={['100%', '677px']}
-          bg={'white'}
-          px={['5vw', '88px']}
-          py={['5vh', '64px']}
-          borderRadius={[0, '16px']}
-          boxShadow={[
-            '',
-            '0px 32px 64px -12px rgba(19, 51, 107, 0.20), 0px 0px 1px 0px rgba(19, 51, 107, 0.20)'
-          ]}
-        >
-          <Box w={['100%', '380px']} flex={'1 0 0'}>
-            {pageType ? (
-              DynamicComponent
-            ) : (
-              <Center w={'full'} h={'full'} position={'relative'}>
-                <Loading fixed={false} />
-              </Center>
-            )}
-          </Box>
-          {feConfigs?.concatMd && (
-            <Box
-              mt={8}
-              color={'primary.700'}
-              fontSize={'mini'}
-              fontWeight={'medium'}
-              cursor={'pointer'}
-              textAlign={'center'}
-              onClick={onOpen}
-            >
-              {t('common:support.user.login.can_not_login')}
-            </Box>
-          )}
-        </Flex>
-
-        {isOpen && <CommunityModal onClose={onClose} />}
+        <LoginContainer
+          onSuccess={loginSuccess}
+          chineseRedirectUrl={ChineseRedirectUrl}
+          languageSelectorPosition="absolute-top-right"
+        />
       </Flex>
-
-      {showRedirect && (
-        <RedirectDrawer
-          isOpen={isOpenRedirect}
-          onClose={onCloseRedirect}
-          onRedirect={() => router.push(ChineseRedirectUrl)}
-          disableDrawer={() => setShowRedirect(false)}
-        />
-      )}
-      {isOpenCookiesDrawer && (
-        <CookiesDrawer
-          onAgree={() => {
-            setLocalCookieVersion(cookieVersion);
-            onCloseCookiesDrawer();
-          }}
-          onClose={onCloseCookiesDrawer}
-        />
-      )}
-    </>
+    </Flex>
   );
 };
-
-function RedirectDrawer({
-  isOpen,
-  onClose,
-  disableDrawer,
-  onRedirect
-}: {
-  isOpen: boolean;
-  onClose: () => void;
-  disableDrawer: () => void;
-  onRedirect: () => void;
-}) {
-  const { t } = useTranslation();
-  return (
-    <Drawer placement="bottom" size={'xs'} isOpen={isOpen} onClose={onClose}>
-      <DrawerOverlay backgroundColor={'rgba(0,0,0,0.2)'} />
-      <DrawerContent py={'1.75rem'} px={'3rem'}>
-        <DrawerCloseButton size={'sm'} />
-        <Flex align={'center'} justify={'space-between'}>
-          <Box>
-            <Box color={'myGray.900'} fontWeight={'500'} fontSize={'1rem'}>
-              {t('login:Chinese_ip_tip')}
-            </Box>
-            <Box
-              color={'primary.700'}
-              fontWeight={'500'}
-              fontSize={'1rem'}
-              textDecorationLine={'underline'}
-              cursor={'pointer'}
-              onClick={disableDrawer}
-            >
-              {t('login:no_remind')}
-            </Box>
-          </Box>
-          <Button ml={'0.75rem'} onClick={onRedirect}>
-            {t('login:redirect')}
-          </Button>
-        </Flex>
-      </DrawerContent>
-    </Drawer>
-  );
-}
-
-function CookiesDrawer({ onClose, onAgree }: { onClose: () => void; onAgree: () => void }) {
-  const { t } = useTranslation();
-
-  return (
-    <Drawer placement="bottom" size={'xs'} isOpen={true} onClose={onClose}>
-      <DrawerOverlay backgroundColor={'rgba(0,0,0,0.2)'} />
-      <DrawerContent py={'1.75rem'} px={'3rem'}>
-        <DrawerCloseButton size={'sm'} />
-        <Flex align={'center'} justify={'space-between'}>
-          <Box>
-            <Box color={'myGray.900'} fontWeight={'500'} fontSize={'1rem'}>
-              {t('login:cookies_tip')}
-            </Box>
-            <Box
-              color={'primary.700'}
-              fontWeight={'500'}
-              fontSize={'1rem'}
-              textDecorationLine={'underline'}
-              cursor={'pointer'}
-              w={'fit-content'}
-              onClick={() => window.open(getDocPath('/docs/agreement/privacy/'), '_blank')}
-            >
-              {t('login:privacy_policy')}
-            </Box>
-          </Box>
-          <Button ml={'0.75rem'} onClick={onAgree}>
-            {t('login:agree')}
-          </Button>
-        </Flex>
-      </DrawerContent>
-    </Drawer>
-  );
-}
 
 export async function getServerSideProps(context: any) {
   return {
